requirements.txt
django-filter>=23.2
pandas>=2.1
jdatetime>=4.1
xlwt>=1.3
dateparser>=1.1
django-money>=3.4
django-phonenumber-field>=7.2.0
django-sortedm2m>=3.0
django-better-admin-arrayfield>=1.4
django-jalali>=6.0